requirements.txt
gunicorn
tensorflow-cpu==2.1.0
flask==1.1.1
numpy==1.17.5
h5py==2.10.0
pillow
